Remove BoardDTO
BoardDTO 제거, getRepresentation을 Board에 추가
