+++ src/test/java/kr/codesquad/freddie/chess/board/BoardTest.java
@@ -91,4 +91,40 @@ class BoardTest {
                 () -> assertThat(board.getBlackPawnsResult()).isEqualTo("PPPPPPPP")
         );
     }
+
+    @Test
+    void getRepresentation_init_board() {
+        String representation = new Board().initialize().getRepresentation();
+
+        String expected = new StringBuilder()
+                .append("........").append(System.lineSeparator())
+                .append("PPPPPPPP").append(System.lineSeparator())
+                .append("........").append(System.lineSeparator())
+                .append("........").append(System.lineSeparator())
+                .append("........").append(System.lineSeparator())
+                .append("........").append(System.lineSeparator())
+                .append("pppppppp").append(System.lineSeparator())
+                .append("........")
+                .toString();
+        assertThat(representation)
+                .isEqualTo(expected);
+    }
+
+    @Test
+    void getRepresentation_not_init_board() {
+        String representation = new Board().getRepresentation();
+
+        String expected = new StringBuilder()
+                .append("........").append(System.lineSeparator())
+                .append("........").append(System.lineSeparator())
+                .append("........").append(System.lineSeparator())
+                .append("........").append(System.lineSeparator())
+                .append("........").append(System.lineSeparator())
+                .append("........").append(System.lineSeparator())
+                .append("........").append(System.lineSeparator())
+                .append("........")
+                .toString();
+        assertThat(representation)
+                .isEqualTo(expected);
+    }
 }
